feedpage update
now it could display the post card, the content of the post card needs to be fixed after post database is set.
test.js is just for refernece DO NOT USE.
the refresh of the feed page needs to be fixed, now it is pull down to refresh, it should be pull up.