sprint 3 userline finished ish
updated the view post page so when commented the user can pull down to refresh and see he/her comment.

changed the my posts button on profile page to my timeline, now it give the user the option to see his/her post, or all of their interactions.

all the server backend done
need to update all the card to make it look pretty